+++ src/collective/messagesviewlet/message.py
@@ -14,8 +14,10 @@ from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from z3c.form.browser.radio import RadioFieldWidget
 from zope import schema
 from zope.interface import alsoProvides
+from zope.interface import implements
 from zope.interface import Invalid
 from zope.interface import invariant
+from zope.schema.fieldproperty import FieldProperty
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 
@@ -156,3 +158,22 @@ def end_index(obj):
         return DateTime(2099, 01, 01)
     else:
         return add_timezone(obj.end)
+
+
+class PseudoMessage(object):
+    """
+        This is not the class used with dexterity !
+        This class is intended to be used in another context to instantiate messages that can be used in viewlet.
+    """
+
+    implements(IMessage)
+
+    # Provide the minimal fields, needed in viewlet templates:
+    msg_type = FieldProperty(IMessage["msg_type"])
+    text = FieldProperty(IMessage["text"])
+    can_hide = FieldProperty(IMessage["can_hide"])
+    hidden_uid = FieldProperty(IMessage["hidden_uid"])
+
+    def __init__(self, **kwargs):
+        for key, value in kwargs.items():
+            setattr(self, key, value)